Move link_text_span.dart out of settings dir
